fix: harden steth reserve yield rollout
