+++ AGENTS.md
@@ -397,7 +397,7 @@ pnpm install
 
 # Indexer
 pnpm indexer:codegen              # Generate types from schema (multichain mainnet)
-pnpm indexer:dev                   # Start indexer (multichain mainnet: Celo + Monad + Ethereum sUSDS)
+pnpm indexer:dev                   # Start indexer (multichain mainnet: Celo + Monad + Ethereum sUSDS/stETH)
 pnpm indexer:mutation              # Targeted StrykerJS baseline for indexer pure logic
 pnpm deploy:indexer                # Push HEAD to envio branch and trigger hosted reindex
 pnpm deploy:indexer:status <commit> --watch  # Wait for registration, then watch sync
@@ -490,7 +490,7 @@ Each package has its own `AGENTS.md` (Claude Code reads them as `CLAUDE.md` via 
 | --------------------- | ---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | -------------------------------------------------------------- |
 | `aegis/`              | NestJS App Engine service polling v2 view calls → Prometheus `/metrics`; also owns the Aegis Grafana dashboard and service-health alert (`aegis/terraform/`)                                                                                                                                                                                                                             | [`aegis/AGENTS.md`](aegis/AGENTS.md)                           |
 | `shared-config/`      | `@mento-protocol/monitoring-config`: chain/token metadata, FX calendar, deployment namespaces. Source of truth — never duplicate chain slugs, explorer URLs, or token labels elsewhere (PR #209). Indexer vendors a copy because Envio builds outside the pnpm workspace.                                                                                                                | [`shared-config/AGENTS.md`](shared-config/AGENTS.md)           |
-| `indexer-envio/`      | Envio HyperIndex (envio@3.0.0): Celo + Monad FPMM + v2 Broker + Ethereum sUSDS reserve-yield accounting. Schema in `schema.graphql`. Handler entry point is `src/EventHandlers.ts` (imports under `src/handlers/`).                                                                                                                                                                      | [`indexer-envio/AGENTS.md`](indexer-envio/AGENTS.md)           |
+| `indexer-envio/`      | Envio HyperIndex (envio@3.0.0): Celo + Monad FPMM + v2 Broker + Ethereum sUSDS/stETH reserve-yield accounting. Schema in `schema.graphql`. Handler entry point is `src/EventHandlers.ts` (imports under `src/handlers/`).                                                                                                                                                                | [`indexer-envio/AGENTS.md`](indexer-envio/AGENTS.md)           |
 | `ui-dashboard/`       | Next.js 16 + Plotly.js + SWR + Tailwind 4. Address book + forensic reports stored in Upstash (`labels` + `reports` hashes), backed up daily to Vercel Blob.                                                                                                                                                                                                                              | [`ui-dashboard/AGENTS.md`](ui-dashboard/AGENTS.md)             |
 | `metrics-bridge/`     | Hasura → Prometheus gauge exporter for v3 alert rules; bounded label cardinality required.                                                                                                                                                                                                                                                                                               | [`metrics-bridge/AGENTS.md`](metrics-bridge/AGENTS.md)         |
 | `integration-probes/` | Quote-only Mento v3 route coverage probes for aggregators and cross-chain routers. Publishes `integration-probes:latest` in Upstash for the dashboard `/integrations` page.                                                                                                                                                                                                              | [`integration-probes/AGENTS.md`](integration-probes/AGENTS.md) |
