+++ indexer-envio/README.md
@@ -1,6 +1,6 @@
 # Mento v3 Envio HyperIndex Indexer
 
-Multichain Envio HyperIndex indexer for Mento v3 — Celo Mainnet (42220), Monad Mainnet (143), and Ethereum Mainnet (1) sUSDS reserve-yield events.
+Multichain Envio HyperIndex indexer for Mento v3 — Celo Mainnet (42220), Monad Mainnet (143), and Ethereum Mainnet (1) sUSDS/stETH reserve-yield events.
 Tracks FPMM pool activity, oracle health, trading limits, rebalancer liveness, and reserve-yield accounting.
 
 ## What It Does
@@ -26,6 +26,7 @@ Listens to on-chain events from Mento v3 contracts and writes structured entitie
 | WormholeNttManager    | `TransferSent`, `TransferRedeemed`, `MessageAttestedTo`, `InboundTransferQueued`                                                                                                                                         |
 | WormholeTransceiver   | `ReceivedMessage`                                                                                                                                                                                                        |
 | Susds                 | `Deposit`, `Withdraw`, `Transfer` (Ethereum only; filtered to tracked Mento reserve wallets)                                                                                                                             |
+| Steth                 | `Transfer` (Ethereum only; filtered to tracked Mento reserve wallets)                                                                                                                                                    |
 
 ### Entities Written
 
@@ -43,7 +44,7 @@ Listens to on-chain events from Mento v3 contracts and writes structured entitie
 | Circuit breakers        | `Breaker`, `BreakerConfig`, `BreakerTripEvent`, `RateFeedDependency`                                                                                                            |
 | Bridge flows            | `BridgeTransfer`, `BridgeAttestation`, `BridgeDailySnapshot`, `BridgeBridger`, `WormholeNttManager`, `WormholeTransferDetail`, `WormholeDestPending`, `WormholeTransferPending` |
 | Volume and participants | `TraderDailySnapshot`, `TraderPoolDailySnapshot`, `AggregatorDailySnapshot`, `VolumeWindowSnapshot`                                                                             |
-| Reserve yield           | `SusdsYieldMovement`, `SusdsCostBasisLot`, `SusdsPosition`, `SusdsYieldSummary`                                                                                                 |
+| Reserve yield           | `SusdsYieldMovement`, `SusdsCostBasisLot`, `SusdsPosition`, `SusdsYieldSummary`, `StethYieldMovement`, `StethCostBasisLot`, `StethPosition`, `StethYieldSummary`                |
 
 ### Pool ID Format
 
@@ -52,11 +53,11 @@ This prevents collisions when the same contract address is deployed on multiple 
 
 ## Configuration
 
-| File                                 | Networks                                                           |
-| ------------------------------------ | ------------------------------------------------------------------ |
-| `config.multichain.mainnet.yaml`     | Celo Mainnet + Monad Mainnet + Ethereum sUSDS (default/production) |
-| `config.multichain.testnet.yaml`     | Celo Sepolia + Monad Testnet                                       |
-| `config.multichain.bridge-only.yaml` | Local bridge-flow validation harness                               |
+| File                                 | Networks                                                                 |
+| ------------------------------------ | ------------------------------------------------------------------------ |
+| `config.multichain.mainnet.yaml`     | Celo Mainnet + Monad Mainnet + Ethereum sUSDS/stETH (default/production) |
+| `config.multichain.testnet.yaml`     | Celo Sepolia + Monad Testnet                                             |
+| `config.multichain.bridge-only.yaml` | Local bridge-flow validation harness                                     |
 
 `config/protocolActors.json` contains manual protocol-controlled caller and
 entry-point overrides for the dashboard volume filter. Pool liquidity-strategy
@@ -91,7 +92,7 @@ GraphQL endpoint: `http://localhost:8080/v1/graphql`
 ### Available Commands (from repo root)
 
 ```bash
-pnpm indexer:codegen                # Generate types (multichain mainnet — Celo + Monad + Ethereum sUSDS)
+pnpm indexer:codegen                # Generate types (multichain mainnet — Celo + Monad + Ethereum sUSDS/stETH)
 pnpm indexer:dev                    # Start local multichain mainnet indexer
 pnpm indexer:testnet:codegen        # Generate types (multichain testnet — Celo Sepolia + Monad testnet)
 pnpm indexer:testnet:dev            # Start local multichain testnet indexer
@@ -226,6 +227,6 @@ See [`STATUS.md`](./STATUS.md) for the static endpoint and deployment reference.
 | `src/EventHandlers.ts`              | Event → entity mapping                                  |
 | `src/helpers.ts`                    | `makePoolId`, `poolIdToAddress` utilities               |
 | `src/rpc.ts` + `src/rpc/`           | RPC read helpers (per-chain clients, effects, fallback) |
-| `config.multichain.mainnet.yaml`    | Production config (Celo + Monad + Ethereum sUSDS)       |
+| `config.multichain.mainnet.yaml`    | Production config (Celo + Monad + Ethereum sUSDS/stETH) |
 | `config/deployment-namespaces.json` | Vendored namespace map for hosted builds                |
 | `abis/`                             | Vendored contract ABI subsets                           |
